Collection#get supports "includeMissing"
Because of the way JS default parameters work, we have to be careful
that options default to a falsy value, because calls only passing one
option by name will have `undefined` for all the others.

Kinto can only store JSON "objects", so if we get a falsy value, it must
be an undefined from IndexedDB. Because we didn't store an
undefined (because we only store JSON "objects"), this must indicate a
missing value.