unclean up warnings
Since we don't transpile to Firefox, and Firefox doesn't support
`for (const ....)` (see
https://bugzilla.mozilla.org/show_bug.cgi?id=1101653), we have to keep
these as `for (let ...)`. The eslint people seem content to not support
this use
case (https://github.com/eslint/eslint/issues/3104#issuecomment-157750198)
despite the fact that they have a dozen WONTFIX bugs about it, so we
can't even silence the warning. Thanks, JS.