Add transactions to Collection layer (fixes #467) (#477)
* collection_test: test updating with numeric IDs

The numeric ID 0 is falsy, so the test fails without this change.

This raises a larger question about whether we even care about/need to
support collections with integer IDs; the Kinto server currently throws
a 503 (see Kinto/kinto#688). For the duration of this branch, I'm
assuming we do.

* collection_test: clear DB before we insert

The beforeEach function is saving records with a different ID
schema. This is going to cause problems for us in the next commit.

* clean up warnings

* define a Collection#execute

This defines a new CollectionTransaction class which will define the
operations we support on transactions. We'll move transactions from
Collection to CollectionTransaction and leave immediate versions (which
are wrapped in their own transaction) in the Collection.

Also, extract getRaw into this transaction class, so we have an initial
example.

* extract Collection#get

* add tests for Collection#execute

These are simple "happy-path" tests; the other tests at the collection
level do a better job of ensuring that everything works.

* extract Collection#deleteAny

Testing this is a little ugly. It's unfortunately not possible to return
the result of anything from inside the transaction, because `deleteAny`
returns the old record and `get` can only access the value of the
record at the beginning of the transaction.

* extract Collection#delete

* add some tests of how transactions should work

* update documentation according to actual behavior

* unclean up warnings

Since we don't transpile to Firefox, and Firefox doesn't support
`for (const ....)` (see
https://bugzilla.mozilla.org/show_bug.cgi?id=1101653), we have to keep
these as `for (let ...)`. The eslint people seem content to not support
this use
case (https://github.com/eslint/eslint/issues/3104#issuecomment-157750198)
despite the fact that they have a dozen WONTFIX bugs about it, so we
can't even silence the warning. Thanks, JS.

* pass the collection into the CollectionTransaction

This lets us use its idSchema.

* extract put and update

These share the _updateRaw function, so it's easier to just extract both
at once.

* use hasOwnProperty to check for missing fields

* fix typo

* extract a small part of create()

This method doesn't really do all that much, since the most significant
part of the Collection#create method is generating a new ID if
necessary, and the execute() function needs to already have the ID
present before it can run the code.

* document transactions

* add more tests about how transactions behave

* refactor `articles` out of transaction tests

* formatting

* minor tweaks to documentation

Thanks @leplatrem for suggesting them.

* get rid of superfluous else

Thanks @leplatrem for the suggestion.