Introduce failing test for updating lastModified (#536)
* Introduce failing test for updating lastModified

I wasn't sure what the correct behavior for manual conflict resolution
was, but I'm pretty sure that we should update lastModified for all of
these cases. So far, we only do so in three of them.

* Collection: Store lastModified at the end of sync()

Doing this here instead of at the end of importChanges means that the
handleConflicts step gets to run before we examine if the sync
succeeded. This fixes the test.

* Add tests to verify stored lastModified

* Reverse order of arguments in test

I think the order is supposed to be "actual" and then "expected".