Remove Firefox storage adapter from this repo (#562)
* [WIP] Move FirefoxAdapter to Gecko code base

Currently, the Firefox storage adapter is versioned in Github. But its unit-tests are in the Gecko codebase.

This pull-request aims to improve that. We provide a build for Gecko that plugs the platform specific dependencies (like UUID, EventEmitter...). And then in the Gecko source code, we'll plug the Firefox adapter in it.

* Rename exported symbol from loadKinto to Kinto

* Make Babel transformations explicit for browsers

* Back to babel development env

* Remove left over

* Add default target for prod, keep development for tests

* Add specific target for noshim