Encode local deletes as remote keeps (#640)
* collection_test: Update tests to desired behavior

pushChanges shouldn't distinguish between `toSync` and `toDelete`
records.

Of course, these tests fail right now.

* collection: pushChanges doesn't distinguish between toDelete and toSync

This still breaks a bunch of tests.

* collection_test: Fix tests to match new gatherLocalChanges

* integration_test: Test transforming local deletes

There are two directions that deletes can be transformed -- local
deletes <-> remote keeps, or local keeps <-> remote deletes. Remote
deletes produce tombstones, which are without much information, so
they're pretty hard to actually make use of in this way. It might be
potentially possible to build something that keeps the record locally
and ignores the tombstone, but 1. this means other clients won't be able
to do anything useful with the record, so it's automatically asymmetric,
and 2. the current approach taken to syncing by kinto.js involves
using the response from the server as the "new" version of the record,
and since the new version of the record is a tombstone, it's impossible
to decode anything useful out of it. So this patch (and this PR) only
focuses on local deletes, for those cases where you want to avoid
leaking information about what you deleted.

Note that getting a transformer to do this correctly is a little tricky,
which is why I wrote an integration test to serve as an example. You
have to make sure your encode() returns a status that isn't "deleted"
but is instead "created" or "updated", and your decode() needs to
somehow recognize the record when it comes back and flag it with
`deleted: true`.

* docs: Document transforming local deletes