integration_test: Refactor tests
There were five copies of this exact stanza. Extract them into a
function.

This requires a bit of finesse because the `tasks` and `syncResult`
variable aren't populated at the time of the function call, but only set
up during each test, so we can't just pass the variables, but rather
closures that provide access to the variables.