integration_test: Add a bunch of tests, some fail
These highlight some of the straight-up errors as well as some of the
misbehaviors that arise when you try to sync with a remote delete.

Note that the behavior for handling resolved conflicts which I believe
to be "correct" requires returning both accepted/rejected versions in
the resolution. This would be a backwards-incompatible change and would
require a major version release.

These tests are really getting unruly. We're up to 2KLOC in just this
file. I think we could probably cut that down a bit by metaprogramming,
but that's outside the scope of this PR.