Fix remote delete (#692)
* integration_test: Refactor tests

There were five copies of this exact stanza. Extract them into a
function.

This requires a bit of finesse because the `tasks` and `syncResult`
variable aren't populated at the time of the function call, but only set
up during each test, so we can't just pass the variables, but rather
closures that provide access to the variables.

* integration_test: Reviewing conflict tests

- Correct descriptions to match tests.
- Ensure lastModified is being handled correctly.

* integration_test: Add a bunch of tests, some fail

These highlight some of the straight-up errors as well as some of the
misbehaviors that arise when you try to sync with a remote delete.

Note that the behavior for handling resolved conflicts which I believe
to be "correct" requires returning both accepted/rejected versions in
the resolution. This would be a backwards-incompatible change and would
require a major version release.

These tests are really getting unruly. We're up to 2KLOC in just this
file. I think we could probably cut that down a bit by metaprogramming,
but that's outside the scope of this PR.

* collection: Don't decode "null"

"null" indicates a remotely deleted record, which doesn't have anything
to decode.

* collection: conflict.remote can be null

* collection: remote=null means delete

* collection: Resolutions now return {accepted, rejected}

This provides more information to calling code and also allows calling
code to "reverse" a resolution if it wants.

* Update docs

* Only test against Kinto 6.0.4 and newer

Kinto 5.3.6 doesn't have PR #1102, which makes it impossible to induce a
conflict on delete. But anyone running such an old Kinto probably ought
to upgrade anyhow.

* @leplatrem review

* collection: Add id and status to resolutions

This way we don't have to check which one succeeded and which one
failed.