Revert 3017452aca9672681c9328ec93853ce472d31faf
It turns out that there's already a Collection#resetSyncStatus, so
IDB#resetSyncStatus just serves to confuse the user.

The place where I intended to use this, I already accidentally used
Collection#resetSyncStatus, so there truly isn't any need for this
method.

The commit that this reverts got squashed as part of #767, so the
commit ID in this commit may not exist in this repository. Instead,
see 7a23ff9b16f3f85dcc92c1d998b2faa86d367006.